+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Development
 
 - DWD Road: Add new station groups, log warning if no data is available, especially if the station group is one of the
   temporarily unavailable ones
+- Explorer: Fix DWD Mosmix request kwargs setup
 
 0.93.0 (06.08.2024)
 *******************